+++ zorka-core/src/main/java/com/jitlogic/zorka/core/spy/TracerLib.java
@@ -356,7 +356,7 @@ public class TracerLib {
      * @throws IOException
      */
     public ZorkaAsyncThread<SymbolicRecord> toZico(String addr, int port, String hostname, String auth,
-                                                   int qlen, int packetSize, int retries, long retryTime, long retryTimeExp,
+                                                   int qlen, long packetSize, int retries, long retryTime, long retryTimeExp,
                                                    int timeout) throws IOException {
         TraceWriter writer = new FressianTraceWriter(symbolRegistry, metricsRegistry);
         ZicoTraceOutput output = new ZicoTraceOutput(writer, addr, port, hostname, auth, qlen, packetSize,
